Fix EditAction
